feat(order, tutorservice): handle edge cases when ordering
- When a learner reserves a service and the order status is updated to
'scheduled':Exclude the reserved date from the availability returned by
the getTutorServiceAvailability method.

- Two learners order the same service at the same time but with
different
total hours:The tutor can choose to accept or reject either order. Once
the tutor accepts one, the other order is automatically rejected.

- A tutor offers multiple services with overlapping dates:
If a learner reserves a service and the order status is 'scheduled'
exclude the reserved date from the availability of the tutor's other
services as well.

- Handling overlapping orders with different durations:
    User 1 orders at 10:30 AM for 5 hours.
    User 2 orders at 1:30 PM for 1 hour.
    Properly handle overlapping availability to prevent conflicts.
    (Completed)